added controllers

+++ VirtualLibrary/virtual-library/src/main/java/virtuallibrary/repositories/BookJpaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import virtuallibrary.models.Book;
 
-public interface BookJpaRepository extends JpaRepository<Book, Integer> {
+public interface BookJpaRepository extends JpaRepository<Book, Long> {
 
 }
