+++ VirtualLibrary/virtual-library/src/main/java/virtuallibrary/repositories/UserJpaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import virtuallibrary.models.User;
 
-public interface UserJpaRepository extends JpaRepository<User, Integer> {
+public interface UserJpaRepository extends JpaRepository<User, Long> {
 
 }
